fix: migrated inbox_page contents to pages/inbox
